Rename module from interface to api
