+++ tests/interface/dealer_test.py
@@ -1,7 +1,7 @@
 from tests.base_unittest import BaseUnitTest
 from tests.record_player import PokerPlayer as RecordMan
 from mock import patch
-from pypokerengine.interface.dealer import Dealer
+from pypokerengine.api.dealer import Dealer
 from pypokerengine.players.sample.fold_man import PokerPlayer as FoldMan
 from pypokerengine.engine.pay_info import PayInfo
